add the yolact_uint8
