+++ source/device/cpu/op/mish/mish_ref.c
@@ -22,6 +22,8 @@
  * Author: 942002795@qq.com
  */
 
+#include "mish_kernel_ref.h"
+
 #include "graph/tensor.h"
 #include "graph/node.h"
 #include "graph/graph.h"
@@ -37,89 +39,6 @@
 #include <math.h>
 
 
-int ref_mish_uint8(struct tensor *input_tensor, struct tensor *output_tensor, int num_thread)
-{
-    int w = input_tensor->dims[3];
-    int h = output_tensor->dims[2];
-    int channels = input_tensor->dims[1];
-    int batch = input_tensor->dims[0];
-
-    int size = h * w;
-    int c_step = h * w;
-    int batch_step = c_step * channels;
-    int total_size = batch_step * batch;
-
-    // dequant
-    uint8_t* input_uint8 = input_tensor->data;
-    uint8_t* output_uint8 = output_tensor->data;
-    float input_scale = input_tensor->scale;
-    float output_scale = output_tensor->scale;
-    int32_t input_zero = input_tensor->zero_point;
-    int32_t output_zero = output_tensor->zero_point;
-
-    float* data_fp32 = sys_malloc(total_size * sizeof(float));
-
-    for(int i = 0; i < total_size; i++)
-        data_fp32[i] = ((float) input_uint8[i] - (float)input_zero) * input_scale;
-
-
-    for (int n = 0; n < batch; n++)
-    {
-#pragma omp parallel for num_threads(num_thread)
-        for (int q = 0; q < channels; q++)
-        {
-            float* src = data_fp32 + batch_step * n + c_step * q;
-            float* dst = data_fp32 + batch_step * n + c_step * q;
-
-            for (int i = 0; i < size; i++)
-            {
-                dst[i] = src[i] * tanhf(log(1 + exp(src[i])));
-            }
-        }
-    }
-
-    // quant
-    for(int i=0; i<total_size; i++)
-    {
-        int udata = round(data_fp32[i] / output_scale + output_zero);
-        if (udata > 255)
-            udata = 255;
-        else if (udata < 0)
-            udata = 0;
-        output_uint8[i] = udata;
-    }
-
-    sys_free(data_fp32);
-
-    return 0;
-}
-
-int ref_mish_fp32(struct tensor* input_tensor, struct tensor* output_tensor, int num_thread)
-{
-    int w = input_tensor->dims[3];
-    int h = output_tensor->dims[2];
-    int channels = input_tensor->dims[1];
-    int size = h * w;
-    int c_step = h * w;
-
-    float* input_data = input_tensor->data;
-    float* out_data = output_tensor->data;
-
-#pragma omp parallel for num_threads(num_thread)
-    for (int q = 0; q < channels; q++)
-    {
-        float* src = input_data + c_step * q;
-        float* dst = out_data + c_step * q;
-
-        for (int i = 0; i < size; i++)
-        {
-            dst[i] = src[i] * tanhf(log(1 + exp(src[i])));
-        }
-    }
-
-    return 0;
-}
-
 static int init_node(struct node_ops* node_ops, struct exec_node* exec_node, struct exec_graph* exec_graph)
 {
     return 0;
@@ -129,7 +48,6 @@ static int release_node(struct node_ops* node_ops, struct exec_node* exec_node, 
 {
     return 0;
 }
-
 
 static int run(struct node_ops* node_ops, struct exec_node* exec_node, struct exec_graph* exec_graph)
 {
